refactor(main): split the HTTP transport out of the manager
`remote-server-errors.ts` holds `RemoteRequestError` and `RemoteProtocolError`. Both are classified by
`instanceof`, so a single definition is load-bearing: a second copy of either class makes every check
against it silently false.

`remote-server-http.ts` holds `requestJson`, `remoteFetch` and `webRtcRequestBody`, plus the
`!response.ok` half of the manager's authenticated fetch, which becomes `throwRemoteResponseError`.
That block and `requestJson` were two hand-written readings of the same failing host response, and
this makes them one. Nothing in the file knows a server exists -- it takes a URL, a token and a
protocol number.

Watched it fail: mutating the non-JSON failure message reddens "treats a non-JSON binary-route
failure as a request error" with `expected ... 'Remote server request failed (502).' but got
'Remote server MUTATED (502).'`.

Two branches of the moved envelope reader turn out to have no test at all -- the rethrow that keeps a
JSON error envelope a request error rather than a protocol error, and a host that claims JSON and
does not send it. Mutating either stays green. Left as found; noted rather than fixed, because this
commit is a move.
